requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
flask-restx==0.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
tomli==2.0.0
typing-extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0
